pyproject.toml
[tool.poetry]
name = "pubmed-cli-tool"
version = "0.1.0"
description = "CLI tool to search PubMed articles with industry affiliations"
authors = ["Kayal <kayalp2001@gmail.com>"]
readme = "README.md"
packages = [{ include = "pubmed_cli_tool", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
click = "^8.1.7"
biopython = "^1.81"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[tool.poetry.scripts]
get-papers-list = "pubmed_cli_tool.cli:search"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
